requirements.txt
google-auth==2.23.3
google-analytics-data==0.7.0
pandas==1.3.3
numpy==1.21.2
openpyxl==3.0.9
